Moved ConfirmCustomAction from the views module to the views module in the admin directory
